Rename histograms script
