+++ figures/__init__.py
@@ -1,4 +1,4 @@
 from . import inputs
 from . import dependencies
-from . import mk_hist
+from . import histograms
 from . import location
