+++ figures/histograms.py
